Migrate AddCalendarDetailsFragment to compose (#40)
