+++ app/src/main/java/at/bitfire/icsdroid/ui/CredentialsComposable.kt
@@ -4,10 +4,6 @@
 
 package at.bitfire.icsdroid.ui
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -24,46 +20,17 @@ import androidx.compose.material.icons.rounded.Visibility
 import androidx.compose.material.icons.rounded.VisibilityOff
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
 import at.bitfire.icsdroid.R
-import at.bitfire.icsdroid.model.CredentialsModel
-
-class CredentialsFragment: Fragment() {
-
-    val model by activityViewModels<CredentialsModel>()
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, inState: Bundle?): View =
-        ComposeView(requireActivity()).apply {
-            setContent {
-                val requiresAuth by model.requiresAuth.observeAsState(false)
-                val username by model.username.observeAsState()
-                val password by model.password.observeAsState()
-
-                LoginCredentialsComposable(
-                    requiresAuth,
-                    username,
-                    password,
-                    onRequiresAuthChange = { model.requiresAuth.postValue(it) },
-                    onUsernameChange = { model.username.postValue(it) },
-                    onPasswordChange = { model.password.postValue(it) },
-                )
-            }
-        }
-
-}
 
 @Composable
 fun LoginCredentialsComposable(
